Fix sort by text fields for elasticsearch, opensearch, typesense and solr (#527)
In elasticsearch and opensearch fields for filtering are postfixed with
`.raw`, it looks the same is required when sorting on them we need to
sort by the raw field which is a keyword instead the normal text field.

Typesense seems to have added sort by text fields in
https://github.com/typesense/typesense/issues/117#issuecomment-1000857228.
In Algolia we have an issue as the uuid needed to be mapped to objectId
and so to filter by it we need change the behaviour.

fixes #523

---------

Co-authored-by: Fernando Saez <fertandil87@gmail.com>