Fix that tests wait for finishing cleanup tasks (#563)
Catched by @Toflar in https://github.com/PHP-CMSIG/search/pull/561.

The deletion in algolia and meilisearch are asynchron and so previous
documents are not correctly waited to be cleanup before running the next
tests. This way it should work again.

This bug effects only our internal tests not users of the library.